+++ inc/gci_yaga_subroutines.h
@@ -7,7 +7,7 @@
 #include <valarray>
 
 using valarray = std::valarray<double>;
-using cand = std::pair<size_t, bool>;
+using cand = std::pair<int, bool>;
 using candvec = std::deque<cand>;
 
 
@@ -21,10 +21,10 @@ double cost_eval(const valarray& lambda, const valarray& costs);
 void cost_calc(const valarray& u, const valarray& gamma, const valarray& norms, const double maxNorm, const cand& r_cand, const cand& q_cand, const cand& p_cand, valarray& cost);
 
 double cost_khi(const double mu, const double sigma, const double y);
-double cost_ncorr(const valarray& u, const size_t r, const size_t p);
-double cost_FN(const valarray& u, const size_t r);
-double cost_ssqe(const valarray& gamma, const size_t r);
-double cost_CP(const valarray& norms, const double maxNorm, const size_t r);
+double cost_ncorr(const valarray& u, const int r, const int p);
+double cost_FN(const valarray& u, const int r);
+double cost_ssqe(const valarray& gamma, const int r);
+double cost_CP(const valarray& norms, const double maxNorm, const int r);
 
 
 #endif // GCI_YAGA_ROUTINES_H
