+++ src/gci_yaga/cost_CP.cpp
@@ -1,7 +1,8 @@
+#include "audio.h"
 #include "gci_yaga_subroutines.h"
 
 
-double cost_CP(const valarray& segmentNorms, const double maxSegmentNorm, const size_t r)
+double cost_CP(const valarray& segmentNorms, const double maxSegmentNorm, const int r)
 {
     return segmentNorms[r] / maxSegmentNorm;
 }
