+++ src/gci_yaga/cost_calc.cpp
@@ -9,11 +9,11 @@ static constexpr double sigma = 0.2;
 
 void cost_calc(const valarray& u, const valarray& gamma, const valarray& norms, const double maxNorm, const cand& r_cand, const cand& q_cand, const cand& p_cand, valarray& cost)
 {
-    size_t r(r_cand.first);
+    int r(r_cand.first);
     bool is_zcr(r_cand.second);
 
-    size_t q(q_cand.first);
-    size_t p(p_cand.first);
+    int q(q_cand.first);
+    int p(p_cand.first);
 
     double FN_r, FN_q;
     double ssqe_r, ssqe_p;
